tests: Upgrade to latest JUnit
In an effort to modernize our often horrible unit tests (yes, horrible -
remember that unit tests are supposed to be the documentation of your
code; they describe what your code is supposed to do), this patch
updates JUnit to the latest version.

The patch doesn't improve anything in itself - we need to utilize the
new features of JUnit to improve things; more patches will follow.

I had to update a few tests because the compiler no longer could figure
out the right type coercions. This is a result of JUnit's Assert class
having additional overloaded versions of assertEquals.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4182/

+++ modules/dcache/src/test/java/org/dcache/pinmanager/PinManagerTests.java
@@ -181,7 +181,7 @@ public class PinManagerTests
         message = processor.messageArrived(message);
 
         assertEquals(0, message.getReturnCode());
-        assertEquals(pin.getPinId(), message.getPinId());
+        assertEquals(pin.getPinId(), (long) message.getPinId());
         assertEquals(pin.getRequestId(), message.getRequestId());
 
         Pin newPin = dao.getPin(pin.getPinId());
@@ -208,7 +208,7 @@ public class PinManagerTests
         message = processor.messageArrived(message);
 
         assertEquals(0, message.getReturnCode());
-        assertEquals(pin.getPinId(), message.getPinId());
+        assertEquals(pin.getPinId(), (long) message.getPinId());
         assertEquals(pin.getRequestId(), message.getRequestId());
 
         Pin newPin = dao.getPin(pin.getPinId());
